Overhaul RDO Manager documentation
The RDO Manager documentation on the website is both
incomplete (it would be hard to get a working deploy
with what is there) and overly in depth (the
architechture section goes into details better handled
by upstream documentation).

Instead, this patch gives instructions to quickly stand
up a virtual deployment, and provides links to upstream
documentation for further reading.