Update quickstart.html.md
I am sharing my own personal experience and also adding a few details to the quickstart page, hoping I can make the experience of installing OpenStack with RDO a breeze. I know it adds a few more lines (which I tried making as easy-reading and enjoyable as possible).

I've recently found the issue regarding /etc/environment which simply blocked my install until I figured out the answer on a forum. Also, the networking part is a bit taken for granted, specially for people starting on the OpenStack (and often "big world computing") scene. The bits about NetworkManager are a known issue for everyone running servers, as well as static IPs. 

I understand that a "quickstart" should be clean and painless, but I also know that a bad start can be frustrating.

I've been fighting the OpenStack beast for the past few weeks now, very unpleased with DevStack and also with the full blown Ubuntu/Autopilot approach (the late sure works, but it was not for me). I am coming back home now, to CentOS, and was pleasantly surprised to see that with a few keystrokes I have a fully working OpenStack system working on my server. Yet, I felt it would be a good idea to contribute, since it worked so well for me, and it could help a lot more people out there. 

Thanks Rich Bowen for the tips.