Reshuffled external connectivity steps to reflect usual order
The arrangement for some steps in the document was not ideal. For
example, we first suggested to create a private network, and only at the
end of the page we noticed that a new user should be created.

With the new arrangement, we now split steps into three sections: first,
all admin invoked commands; then create a new user; then create tenant
resources as this user.