Update packstack min RAM requirements to match real world
The docs suggest packstack can be run in a machine with 4 GB
of RAM. While this may have worked in the past, it does not
work any longer. Even installing in a machine with 8 GB triggers
out of memory failures. A 12 GB machine was able to complete the
all-in-one install successfully, but the services almost
immediately start getting reaped by the OOM-killer. Thus document
that 16 GB of RAM is required for packstack.

Signed-off-by: Daniel P. Berrange <berrange@redhat.com>